renamed utils to bnn_utils to be clearer
